+++ videos/videos.py
@@ -4,13 +4,13 @@ from flask import Blueprint, render_template, request, flash, redirect, url_for
 from videos.video_entry import VideoEntry
 from validations import validate_url
 from setup import db
-from download import download
+from videos.download import download
 from flask_login import current_user, login_required
 from forms import VideoEditForm
-from typing import Type, Optional
+from typing import Type, Optional, Union
 videos = Blueprint('videos', __name__)
 
-def __add_video_to_db() -> Optional[t.Type["BaseResponse"]]:
+def __add_video_to_db() -> Optional[Type["BaseResponse"]]:
     url = request.form.get('url')
     if not validate_url(url):
         flash('You must enter a youtube url', category='error')
@@ -40,7 +40,7 @@ def add_video() -> str:
     return render_template('add_video.html')
 
 @videos.route('/edit/', methods=['GET', 'POST'])
-def edit() -> Union[str, t.Type["BaseResponse"]]:
+def edit() -> Union[str, Type["BaseResponse"]]:
     video = VideoEntry.query.get(int(request.args.get('id')))
     print(request.form)
     if request.method == 'POST':
@@ -51,7 +51,7 @@ def edit() -> Union[str, t.Type["BaseResponse"]]:
 
     return render_template('edit_video.html', form=VideoEditForm(), video=video)
 @videos.route('/download/', methods=['GET', 'POST'])
-def download_video() -> t.Type["BaseResponse"]:
+def download_video() -> Type["BaseResponse"]:
     url = request.args.get('url')
     download(url)
     videos_with_given_url = VideoEntry.query.filter_by(url=url)
@@ -62,7 +62,7 @@ def download_video() -> t.Type["BaseResponse"]:
 
 @videos.route('/delete/')
 @login_required
-def delete_video() -> t.Type["BaseResponse"]:
+def delete_video() -> Type["BaseResponse"]:
     video_id = int(request.args.get('id'))
     video = VideoEntry.query.get(video_id)
     if video:
